+++ get-opt/src/defines.hpp
@@ -1,0 +1,8 @@
+#pragma once
+
+#include <string>
+#include <vector>
+
+
+typedef std::string Arg;
+typedef std::vector<Arg> Args;
