+++ get-opt/src/terminal.hpp
@@ -1,0 +1,36 @@
+#pragma once
+
+#include <map>
+#include <iostream>
+#include "defines.hpp"
+#include "key.hpp"
+
+
+class Listener {
+public:
+	virtual Listener* getCopy() = 0;
+	virtual void execute(Args) = 0;
+};
+
+class Terminal {
+private:
+	std::map<Key, Listener*> data;
+
+public:
+	Terminal();
+	Terminal(const Terminal& term);
+	~Terminal();
+
+	Terminal& operator=(const Terminal& term);
+
+	void addKey(Key key, Listener* lnr = nullptr);
+	void delKey(Key key);
+
+	void attachKey(Key key, Listener* lnr);
+	void detachKey(Key key);
+
+	void attachRoot(Listener* lnr);
+	void detachRoot();
+	
+	void execute(Args input);
+};
