+++ get-opt/src/utils.hpp
@@ -1,0 +1,9 @@
+#pragma once
+
+#include <string>
+#include <exception>
+#include <stdexcept>
+
+
+double str_num(std::string str);
+std::string num_str(double num);
